+++ game/src/enemy2.lua
@@ -6,40 +6,44 @@ function Enemy2:new(path)
     self.imageQuad = Images.enemy2
     self.width = self.imageQuad.width
     self.height = self.imageQuad.height
+    self.health = 6
+    self.score = 20
 end
 
-function Enemy2:fire(newBullets)
+function Enemy2:fire()
     local startX = self.path.x + (self.width - Images.enemybullet.width) / 2
     local startY = self.path.y + self.height
     local speed = 100 + love.math.random(100)
-    table.insert(newBullets, {
-        x = startX,
-        y = startY,
-        speed = speed,
-        angle = 36,
-        colour = self.colour
-    })
-    table.insert(newBullets, {
-        x = startX,
-        y = startY,
-        speed = speed,
-        angle = 72,
-        colour = self.colour
-    })
-    table.insert(newBullets, {
-        x = startX,
-        y = startY,
-        speed = speed,
-        angle = 108,
-        colour = self.colour
-    })
-    table.insert(newBullets, {
-        x = startX,
-        y = startY,
-        speed = speed,
-        angle = 144,
-        colour = self.colour
-    })
+    return {
+        {
+            x = startX,
+            y = startY,
+            speed = speed,
+            angle = 36,
+            colour = self.colour
+        },
+        {
+            x = startX,
+            y = startY,
+            speed = speed,
+            angle = 72,
+            colour = self.colour
+        },
+        {
+            x = startX,
+            y = startY,
+            speed = speed,
+            angle = 108,
+            colour = self.colour
+        },
+        {
+            x = startX,
+            y = startY,
+            speed = speed,
+            angle = 144,
+            colour = self.colour
+        }
+    }
 end
 
 function Enemy2:draw()